requirements.txt
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
scikit-learn==1.4.2
statsmodels==0.14.2
xgboost==2.0.3

# IMPORTANT — compatible with torch 2.2 CPU
pytorch-lightning==2.2.5
torchmetrics==1.3.2

optuna==3.6.1
wandb==0.17.0
pyyaml==6.0.1
seaborn==0.13.2
tqdm==4.66.4
python-dotenv==1.0.1
